refactor: Split StepPlayer into CodingStep and FollowingStep
